Enforced column names to be at most 63 bytes
Replaced usage of &'static str with new ColumnName type.
This new type stores its bytes inline and can be manipulated in const fns.

This is not really a restriction because postgres has a limit of 63 anyway.
(Msql and mariadb limit column names to 64; I couldn't find a limit for sqlite)